# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateUserRequest:
    properties:
      deptIds:
        description: 部门id数组
        example:
        - '["UUID"]'
        items:
          type: string
        type: array
      desc:
        description: 描述
        example: 普通用户
        type: string
      leaderUserId:
        description: 直属上级用户ID
        example: UUID
        type: string
      password:
        description: 密码
        example: "123456"
        type: string
      phone:
        description: 手机号
        example: "13800138000"
        type: string
      realName:
        description: 真实姓名
        example: 新用户
        type: string
      roleIds:
        description: 角色id数组
        example:
        - '["UUID"]'
        items:
          type: string
        type: array
      username:
        description: 登录用户名
        example: newuser
        type: string
    required:
    - password
    - realName
    - username
    type: object
...
  models.ProfileResponse:
    properties:
      avatar:
        description: 用户头像URL
        example: https://xxx/avatar.jpg
        type: string
      createDate:
        description: 创建时间
        example: "2024-01-01 12:00:00"
        type: string
      deptIds:
        description: 部门id数组
        items:
          type: string
        type: array
      deptTitles:
        description: 部门名称数组
        items:
          type: string
        type: array
      desc:
        description: 用户描述
        example: 超级管理员
        type: string
      email:
        description: 邮箱
        example: admin@example.com
        type: string
      homePath:
        description: 首页路径
        example: /dashboard/analytics
        type: string
      leaderUserId:
        description: 直属上级用户ID
        example: UUID
        type: string
      leaderUserName:
        description: 直属上级用户姓名
        example: 张三
        type: string
      phone:
        description: 手机号
        example: "13800138000"
        type: string
      realName:
        description: 真实姓名
        example: 管理员
        type: string
      roleIds:
        description: 角色id数组
        items:
          type: string
        type: array
      roleTitles:
        description: 角色名称数组
        items:
          type: string
        type: array
      status:
        description: 用户状态 0=禁用 1=启用
        example: 1
        type: integer
      updateDate:
        description: 更新时间
        example: "2024-01-01 12:00:00"
        type: string
      userId:
        description: 用户ID
        example: UUID
        type: string
      username:
        description: 登录用户名
        example: admin
        type: string
    type: object
...
  models.Response:
    properties:
      code:
        description: 状态码 0=成功 -1=失败
        example: 0
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息
      message:
        description: 提示信息
        example: ok
        type: string
    type: object
...
  models.UpdateUserRequest:
    properties:
      deptIds:
        description: 部门id数组
        example:
        - '["UUID"]'
        items:
          type: string
        type: array
      desc:
        description: 描述
        example: 普通用户
        type: string
      leaderUserId:
        description: 直属上级用户ID
        example: UUID
        type: string
      phone:
        description: 手机号
        example: "13800138000"
        type: string
      realName:
        description: 真实姓名
        example: 新用户
        type: string
      roleIds:
        description: 角色id数组
        example:
        - '["UUID"]'
        items:
          type: string
        type: array
      username:
        description: 登录用户名
        example: newuser
        type: string
    required:
    - realName
    - username
    type: object
...
  utils.PageResult:
    properties:
      items: {}
      total:
        type: integer
    type: object
...
host: localhost:9191
info:
  contact: {}
  description: Hive Admin Go 后端 API 接口文档
  title: Hive Admin API
  version: "1.0"
paths:
...
  /system/users:
    delete:
      consumes:
      - application/json
      description: 批量删除用户
      parameters:
      - description: 用户ID列表
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 系统管理-用户管理
    get:
      consumes:
      - application/json
      description: 分页获取用户列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 用户姓名
        in: query
        name: realName
        type: string
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      - description: 部门ID，查询该部门及子部门的用户
        in: query
        name: deptId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/utils.PageResult'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/models.ProfileResponse'
                        type: array
                    type: object
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户列表
      tags:
      - 系统管理-用户管理
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建用户
      tags:
      - 系统管理-用户管理
  /system/users/{userId}:
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户详情
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ProfileResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户详情
      tags:
      - 系统管理-用户管理
    put:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: string
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新用户
      tags:
      - 系统管理-用户管理
  /system/users/{userId}/status:
    put:
      consumes:
      - application/json
      description: 更新用户启用/禁用状态
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: string
      - description: 状态
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新用户状态
      tags:
      - 系统管理-用户管理
  /system/users/all:
    get:
      consumes:
      - application/json
      description: 获取所有用户（不分页）
      parameters:
      - description: 用户姓名
        in: query
        name: realName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ProfileResponse'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取所有用户
      tags:
      - 系统管理-用户管理
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"